Remove vendor and fix code generation
Mimic the `update-codegen.sh` implementation from Flagger, this
implementation handles modules correctly.

The controlle code-generator is not compatible with modules and vendor,
so this removes the vendor folder and all of the go flags that set the
vendor mode.

Update to client-go 0.21.3, which is the latest version with v1beta1
networking support.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>